requirements.txt
telethon
curl_cffi
beautifulsoup4
upstash-redis
python-dotenv
flask
